refactor: sse module 분리, 라우팅 버그, hydration 버그 해결
