+++ src/features/session/components/SessionDetailModal/SessionDetailModal.tsx
@@ -2,14 +2,13 @@
 
 import { useEffect, useState } from "react";
 
-import { useRouter } from "next/navigation";
-
 import { Button } from "@/components/Button/Button";
 import { ButtonLink } from "@/components/Button/ButtonLink";
 import { CloseIcon } from "@/components/Icon/CloseIcon";
 import { SessionJoinModal } from "@/features/lobby/components/SessionJoinModal";
 import { useIsAuthenticated, useMe } from "@/features/member/hooks/useMemberHooks";
 import { useDialog } from "@/hooks/useDialog";
+import { navigateWithHardReload } from "@/lib/navigation/hardNavigate";
 
 import { useSessionDetail, useWaitingRoom } from "../../hooks/useSessionHooks";
 import { Card } from "../Card/Card";
@@ -21,7 +20,6 @@ interface SessionDetailModalProps {
 }
 
 export function SessionDetailModal({ sessionId }: SessionDetailModalProps) {
-  const router = useRouter();
   const { dialogRef, handleClose, handleBackdropClick } = useDialog("/");
   const { data } = useSessionDetail(sessionId);
   const isAuthenticated = useIsAuthenticated();
@@ -49,15 +47,16 @@ export function SessionDetailModal({ sessionId }: SessionDetailModalProps) {
   const isOngoingStatus = session?.status === "진행중" || session?.status === "진행 중";
 
   // 자동 리다이렉트: 이미 참여 중이면 해당 세션 페이지로 이동
+  // Intercepting Route(@modal) 컨텍스트에서 벗어나기 위해 하드 네비게이션 사용
   useEffect(() => {
     if (isParticipant && isWaitingStatus) {
       dialogRef.current?.close();
-      router.replace(`/session/${sessionId}/waiting`);
+      navigateWithHardReload(`/session/${sessionId}/waiting`);
     } else if (isParticipant && isOngoingStatus) {
       dialogRef.current?.close();
-      router.replace(`/session/${sessionId}`);
+      navigateWithHardReload(`/session/${sessionId}`);
     }
-  }, [isParticipant, isWaitingStatus, isOngoingStatus, sessionId, router, dialogRef]);
+  }, [isParticipant, isWaitingStatus, isOngoingStatus, sessionId, dialogRef]);
 
   // 참여 여부 확인 중인지 여부
   const isCheckingParticipation = isAuthenticated && isWaitingRoomLoading;
